backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
opencv-python==4.10.0.84
mediapipe==0.10.14
numpy==1.26.4
Pillow==10.4.0